+++ visualizer/src/ships/mod.rs
@@ -1,21 +1,27 @@
-use bevy::prelude::*;
+use bevy::{prelude::*, time::common_conditions::on_timer};
+use bevy_rand::{prelude::WyRand, resource::GlobalEntropy};
+use rand::Rng;
+use std::time::Duration;
 
-use crate::map::{Fuel, Position};
+use crate::map::{Fuel, Position, ShipIdentity};
+use hud::{HudState, render_hud};
+
+mod hud;
 
 const TILE_SIZE: u32 = 64;
 
 #[derive(Resource)]
-pub struct Material {
+pub struct ShipMaterial {
     texture: Handle<Image>,
     layout: Handle<TextureAtlasLayout>,
 }
 
-impl FromWorld for Material {
+impl FromWorld for ShipMaterial {
     fn from_world(world: &mut World) -> Self {
         let texture: Handle<Image> = world
             .get_resource::<AssetServer>()
             .unwrap()
-            .load("asteroid/sprite.png");
+            .load("ship/sprite.png");
 
         let layout = world
             .get_resource_mut::<Assets<TextureAtlasLayout>>()
@@ -33,14 +39,20 @@ impl FromWorld for Material {
 }
 
 #[derive(Bundle)]
-pub struct Asteroid {
+pub struct Ship {
     sprite_sheet: SpriteSheetBundle,
+    identity: ShipIdentity,
     position: Position,
     fuel: Fuel,
 }
 
-impl Asteroid {
-    pub fn new(position: Position, fuel: Fuel, material: &Material) -> Self {
+impl Ship {
+    pub fn new(
+        identity: ShipIdentity,
+        position: Position,
+        fuel: Fuel,
+        material: &ShipMaterial,
+    ) -> Self {
         Self {
             sprite_sheet: SpriteSheetBundle {
                 atlas: TextureAtlas {
@@ -48,12 +60,13 @@ impl Asteroid {
                     index: 0,
                 },
                 transform: Transform {
-                    scale: Vec3::new(0.8, 0.8, 1.0),
+                    scale: Vec3::new(1.5, 1.5, 1.0),
                     ..Default::default()
                 },
                 texture: material.texture.clone(),
                 ..Default::default()
             },
+            identity,
             position,
             fuel,
         }
@@ -70,10 +83,30 @@ fn render(mut query: Query<(&mut Transform, &Position)>) {
     }
 }
 
-pub struct AsteroidPlugin;
+fn random_move(
+    mut query: Query<(&mut Position, &Fuel), With<ShipIdentity>>,
+    mut rng: ResMut<GlobalEntropy<WyRand>>,
+) {
+    for (mut pos, fuel) in query.iter_mut() {
+        if fuel.available > 300 {
+            pos.x += rng.gen_range(-1..1);
+            pos.y += rng.gen_range(-1..1);
+        }
+    }
+}
+pub struct ShipsPlugin;
 
-impl Plugin for AsteroidPlugin {
+impl Plugin for ShipsPlugin {
     fn build(&self, app: &mut App) {
-        app.init_resource::<Material>().add_systems(Update, render);
+        app.init_resource::<ShipMaterial>()
+            .add_systems(
+                Update,
+                (
+                    (render).chain(),
+                    random_move.run_if(on_timer(Duration::from_secs(2))),
+                    render_hud,
+                ),
+            )
+            .insert_state(HudState(None));
     }
 }
