+++ visualizer/src/ui.rs
@@ -1,0 +1,3 @@
+use bevy::prelude::*;
+
+pub const PRIMARY_COLOR: Color = Color::rgb(255., 255., 0.);
